+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-v0.3.0 (in development)
------------------------
+v0.3.0 (2021-06-23)
+-------------------
 - Add `filter` and `exclude` arguments
 - Add `SelectNames`, `SelectGlob`, `SelectRegex`, `SELECT_DOTS`, `SELECT_VCS`,
   `SELECT_VCS_DIRS`, and `SELECT_VCS_FILES` for easy filtering by entry name